refactor(member): Member 도메인, 엔티티 클래스 통합
- Jdbc기반 레포지토리, 헬퍼 클래스 삭제
- 기존의 도메인, 엔티티 분리 구조에서 Member클래스를 도메인, 엔티티로 사용
- 변경 사항 반영하여 Service로직 수정
