+++ src/main/java/gift/member/domain/Member.java
@@ -1,32 +1,40 @@
 package gift.member.domain;
 
-import org.springframework.util.Assert;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
 
-public record Member(
-    Long id,
-    String name
-) {
+@Entity
+@Table(name = "member")
+public class Member {
 
-    public Member {
-        validateName(name);
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(nullable = false)
+    private String name;
+
+    protected Member() {
+    }
+
+    private Member(Long id, String name) {
+        this.id = id;
+        this.name = name;
     }
 
     public static Member of(String name) {
         return new Member(null, name);
     }
 
-    public static Member withId(Long id, String name) {
-        validateId(id);
-        return new Member(id, name);
+    public Long getId() {
+        return id;
     }
 
-    private static void validateId(Long id) {
-        if (id == null || id < 0) {
-            throw new IllegalArgumentException("id값은 null이거나 음수일 수 없습니다.");
-        }
-    }
-
-    private static void validateName(String name) {
-        Assert.hasText(name, "name은 null이거나 빈 값일 수 없습니다.");
+    public String getName() {
+        return name;
     }
 }
