--- src/main/java/gift/member/repository/JpaMemberRepository.java
@@ -1,8 +1,0 @@
-package gift.member.repository;
-
-import gift.member.jpa.MemberEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface JpaMemberRepository extends JpaRepository<MemberEntity, Long> {
-
-}
