+++ src/main/java/gift/member/repository/MemberJpaRepository.java
@@ -1,0 +1,8 @@
+package gift.member.repository;
+
+import gift.member.domain.Member;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MemberJpaRepository extends JpaRepository<Member, Long> {
+
+}
