+++ src/main/java/gift/member/service/MemberService.java
@@ -2,15 +2,15 @@ package gift.member.service;
 
 import gift.member.domain.Member;
 import gift.member.exception.MemberNotFoundException;
-import gift.member.repository.MemberRepository;
+import gift.member.repository.MemberJpaRepository;
 import org.springframework.stereotype.Service;
 
 @Service
 public class MemberService {
 
-    private final MemberRepository memberRepository;
+    private final MemberJpaRepository memberRepository;
 
-    public MemberService(MemberRepository memberRepository) {
+    public MemberService(MemberJpaRepository memberRepository) {
         this.memberRepository = memberRepository;
     }
